@@ -12,17 +12,17 @@
   "private": true,
   "dependencies": {
     "@angular-material-components/datetime-picker": "^7.0.1",
-    "@angular/animations": "~13.2.0",
+    "@angular/animations": "^14.2.12",
     "@angular/cdk": "^13.2.1",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
-    "@angular/forms": "~13.2.0",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
     "@angular/material": "^13.2.1",
     "@angular/material-luxon-adapter": "^13.3.3",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@danielmoncada/angular-datetime-picker": "^13.1.1",
     "angular-imask": "^6.4.1-alpha.0",
     "bootstrap": "^5.1.3",
@@ -38,10 +38,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.10",
-    "@angular/cli": "~13.2.2",
-    "@angular/compiler-cli": "~13.2.0",
-    "@angular/localize": "^13.2.2",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/localize": "^14.2.12",
     "@types/jasmine": "~3.10.0",
     "@types/luxon": "^2.3.1",
     "@types/node": "^12.11.1",
@@ -54,6 +54,6 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "postcss": "^8.4.7",
     "tailwindcss": "^3.0.23",
-    "typescript": "~4.5.2"
+    "typescript": "~4.8.4"
   }
 }
