@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { LuxonDateAdapter, MAT_LUXON_DATE_ADAPTER_OPTIONS } from '@angular/material-luxon-adapter';
 import { DateAdapter, MAT_DATE_FORMATS, MAT_DATE_LOCALE } from '@angular/material/core';
 import { MatDatepicker } from '@angular/material/datepicker';
@@ -31,7 +31,7 @@ export class YearPickerComponent implements OnInit {
   // `placeholderText` Input from datetime main component
   @Input() placeholderText!: any;
 
-  @Input() fieldControl!: FormControl;
+  @Input() fieldControl!: UntypedFormControl;
 
   result!: string;
 
